Settings screen: improve discoverability of save button
