+++ assets/src/settings-page/settings-footer.js
@@ -1,15 +1,50 @@
 /**
+ * External dependencies
+ */
+import PropTypes from 'prop-types';
+
+/**
  * WordPress dependencies
  */
 import { Button } from '@wordpress/components';
-import { useContext } from '@wordpress/element';
+import { useContext, useState, useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
+/**
+ * Internal dependencies
+ */
+import { AMPNotice, NOTICE_TYPE_ERROR, NOTICE_TYPE_SUCCESS } from '../components/amp-notice';
 
 /**
  * Internal dependencies
  */
 import { Options } from '../components/options-context-provider';
 import { ReaderThemes } from '../components/reader-themes-context-provider';
+import { ErrorContext } from '../components/error-context-provider';
+
+/**
+ * Renders an error notice.
+ *
+ * @param {Object} props Component props.
+ * @param {string} props.errorMessage Error message text.
+ */
+function ErrorNotice( { errorMessage } ) {
+	return (
+		<div className="amp-error-notice">
+			<AMPNotice type={ NOTICE_TYPE_ERROR }>
+				<p>
+					<strong>
+						{ __( 'Error:', 'amp' ) }
+					</strong>
+					{ ' ' }
+					{ errorMessage }
+				</p>
+			</AMPNotice>
+		</div>
+	);
+}
+ErrorNotice.propTypes = {
+	errorMessage: PropTypes.string,
+};
 
 /**
  * The bottom section of the settings page.
@@ -17,6 +52,44 @@ import { ReaderThemes } from '../components/reader-themes-context-provider';
 export function SettingsFooter() {
 	const { didSaveOptions, editedOptions, hasOptionsChanges, savingOptions } = useContext( Options );
 	const { downloadingTheme } = useContext( ReaderThemes );
+	const { error } = useContext( ErrorContext );
+
+	const [ savedNoticeClass, setSavedNoticeClass ] = useState( '' );
+
+	/**
+	 * Show the success notice after options have saved.
+	 */
+	useEffect( () => {
+		if ( didSaveOptions && ! downloadingTheme ) {
+			setSavedNoticeClass( 'visible' );
+		}
+	}, [ didSaveOptions, downloadingTheme ] );
+
+	/**
+	 * If the success notice is showing and updates have been made, hide the notice.
+	 */
+	useEffect( () => {
+		if ( 'visible' === savedNoticeClass && hasOptionsChanges ) {
+			setSavedNoticeClass( 'dismissed' );
+		}
+	}, [ savedNoticeClass, hasOptionsChanges ] );
+
+	/**
+	 * Hide the success notice after several seconds.
+	 */
+	useEffect( () => {
+		if ( 'visible' === savedNoticeClass ) {
+			const timeout = setTimeout( () => {
+				setSavedNoticeClass( 'dismissed' );
+			}, 9000 );
+
+			return () => {
+				clearTimeout( timeout );
+			};
+		}
+
+		return () => undefined;
+	}, [ savedNoticeClass ] );
 
 	const { reader_theme: readerTheme, theme_support: themeSupport } = editedOptions;
 
@@ -30,11 +103,23 @@ export function SettingsFooter() {
 	return (
 		<section className="settings-footer">
 			<Button isPrimary disabled={ disabled } isBusy={ savingOptions || downloadingTheme } type="submit">
-				{ __( 'Save changes', 'amp' ) }
+				{ savingOptions || downloadingTheme ? __( 'Saving', 'amp' ) : __( 'Save', 'amp' ) }
 				<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 64 64">
 					<path d="M43.16 10.18c-0.881-0.881-2.322-0.881-3.203 0s-0.881 2.322 0 3.203l16.335 16.335h-54.051c-1.281 0-2.242 1.041-2.242 2.242 0 1.281 0.961 2.322 2.242 2.322h54.051l-16.415 16.335c-0.881 0.881-0.881 2.322 0 3.203s2.322 0.881 3.203 0l20.259-20.259c0.881-0.881 0.881-2.322 0-3.203l-20.179-20.179z" />
 				</svg>
 			</Button>
+			{ error && <ErrorNotice errorMessage={ error.message || __( 'An error occurred. You might be offline or logged out.', 'amp' ) } /> }
+			{ ! error && ! hasOptionsChanges && didSaveOptions && ! downloadingTheme && (
+				<AMPNotice
+					className={ `amp-save-success-notice ${ savedNoticeClass }` }
+					type={ NOTICE_TYPE_SUCCESS }
+					aria-hidden={ 'visible' !== savedNoticeClass }
+				>
+					<p>
+						{ __( 'Saved', 'amp' ) }
+					</p>
+				</AMPNotice>
+			) }
 		</section>
 	);
 }
