+++ assets/src/settings-page/style.css
@@ -20,7 +20,7 @@ html {
 
 #amp-settings {
 	max-width: 1060px;
-	margin: 0 auto;
+	margin: 0 auto 7.5rem;
 }
 
 #amp-settings > h1 {
@@ -297,71 +297,54 @@ li.error-kept {
 	}
 }
 
-@keyframes slideIn {
-
-	from {
-		transform: translate3d(300px, 0, 0);
-	}
-
-	to {
-		transform: translate3d(0, 0, 0);
-	}
-}
-
-@keyframes slideOut {
-
-	from {
-		transform: translate3d(0, 0, 0);
-	}
-
-	to {
-		transform: translate3d(300px, 0, 0);
-	}
-}
-
 .amp .amp-save-success-notice.amp-notice:not(.dismissed):not(.visible) {
 	transform: translate3d(300px, 0, 0);
-}
-
-.amp .amp-save-success-notice.amp-notice.visible {
-	animation: slideIn;
-	animation-delay: 0s;
-	animation-duration: 0.5s;
-	animation-fill-mode: both;
-}
-
-.amp .amp-save-success-notice.amp-notice.dismissed {
-	animation: slideOut;
-	animation-delay: 0s;
-	animation-duration: 0.5s;
-	animation-fill-mode: both;
 }
 
 .amp .amp-save-success-notice.amp-notice,
 .amp-error-notice .amp-notice {
 	bottom: 3rem;
-	box-shadow: 0 30px 20px rgba(0, 0, 0, 0.1);
-	max-width: 300px;
+	padding-left: 1.5rem;
 	padding-right: 3rem;
-	position: fixed;
-	right: -1rem;
-	width: 100%;
 	z-index: 99;
 }
 
-.amp-error-notice .amp-notice {
-	animation: slideIn;
-	animation-delay: 0s;
-	animation-duration: 0.5s;
-	animation-fill-mode: both;
-	max-width: 400px;
-}
-
 .settings-footer {
-	padding-bottom: 2.5rem;
-	padding-top: 2.5rem;
+	align-items: center;
+	background: var(--amp-settings-color-background);
+	bottom: 0;
+	box-shadow: 0 -5px 15px rgba(0, 0, 0, 0.05);
+	display: flex;
+	justify-content: space-between;
+	left: 0;
+	padding: 2.5rem;
+	position: fixed;
+	right: 0;
+	z-index: 20;
 }
 
+.settings-footer .components-button.is-primary {
+	box-shadow: none;
+}
+
+@media (min-width: 783px) {
+
+	.settings-footer {
+		left: 160px;
+	}
+
+	.wp-admin.folded .settings-footer,
+	.wp-admin.auto-fold .settings-footer {
+		left: 36px;
+	}
+}
+
+@media (min-width: 961px) {
+
+	.wp-admin.auto-fold .settings-footer {
+		left: 160px;
+	}
+}
 
 #template-mode-reader-container {
 	margin-bottom: 0;
